fpu: explicit integer casts on the banker's-rounding ceil/floor results
Backport from upstream mupen64plus-core: round_l_s,
round_w_s, round_l_d and round_w_d each assign the result
of ceilf/floorf/ceil/floor (which return float/double) to
an int64_t* or int32_t* destination through a conditional
expression. Implicit float-to-integer narrowing in this
context produces a stream of warnings on MSVC and modern
clang (-Wfloat-conversion / C4244), and the warnings get
duplicated for every translation unit that includes
fpu.h.

The arithmetic result is identical -- the implicit
conversion already truncates the same way an explicit cast
does -- so this is a pure warning-suppression change.
Cast each ceil/floor result in the if (remainder == 0.5)
branches of all four functions to the destination integer
type. The other-branch (int64_t)/(int32_t) casts on the
roundf/round call already match the destination, so they
need no change.

+++ mupen64plus-core/src/r4300/fpu.h
@@ -117,9 +117,9 @@ M64P_FPU_INLINE void round_l_s(const float *source,int64_t *dest)
   if (remainder == 0.5)
   {
     if (*source < 0)
-      *dest = (int64_t)truncf(*source) % 2 != 0 ? floorf(*source) : ceilf(*source);
+      *dest = (int64_t)truncf(*source) % 2 != 0 ? (int64_t)floorf(*source) : (int64_t)ceilf(*source);
     else
-      *dest = (int64_t)truncf(*source) % 2 != 0 ? ceilf(*source) : floorf(*source);
+      *dest = (int64_t)truncf(*source) % 2 != 0 ? (int64_t)ceilf(*source) : (int64_t)floorf(*source);
   }
   else
     *dest = (int64_t) roundf(*source);
@@ -130,9 +130,9 @@ M64P_FPU_INLINE void round_w_s(const float *source,int32_t *dest)
   if (remainder == 0.5)
   {
     if (*source < 0)
-      *dest = (int32_t)truncf(*source) % 2 != 0 ? floorf(*source) : ceilf(*source);
+      *dest = (int32_t)truncf(*source) % 2 != 0 ? (int32_t)floorf(*source) : (int32_t)ceilf(*source);
     else
-      *dest = (int32_t)truncf(*source) % 2 != 0 ? ceilf(*source) : floorf(*source);
+      *dest = (int32_t)truncf(*source) % 2 != 0 ? (int32_t)ceilf(*source) : (int32_t)floorf(*source);
   }
   else
     *dest = (int32_t) roundf(*source);
@@ -168,9 +168,9 @@ M64P_FPU_INLINE void round_l_d(const double *source,int64_t *dest)
   if (remainder == 0.5)
   {
     if (*source < 0)
-      *dest = (int64_t)trunc(*source) % 2 != 0 ? floor(*source) : ceil(*source);
+      *dest = (int64_t)trunc(*source) % 2 != 0 ? (int64_t)floor(*source) : (int64_t)ceil(*source);
     else
-      *dest = (int64_t)trunc(*source) % 2 != 0 ? ceil(*source) : floor(*source);
+      *dest = (int64_t)trunc(*source) % 2 != 0 ? (int64_t)ceil(*source) : (int64_t)floor(*source);
   }
   else
     *dest = (int64_t) round(*source);
@@ -181,9 +181,9 @@ M64P_FPU_INLINE void round_w_d(const double *source,int32_t *dest)
   if (remainder == 0.5)
   {
     if (*source < 0)
-      *dest = (int32_t)trunc(*source) % 2 != 0 ? floor(*source) : ceil(*source);
+      *dest = (int32_t)trunc(*source) % 2 != 0 ? (int32_t)floor(*source) : (int32_t)ceil(*source);
     else
-      *dest = (int32_t)trunc(*source) % 2 != 0 ? ceil(*source) : floor(*source);
+      *dest = (int32_t)trunc(*source) % 2 != 0 ? (int32_t)ceil(*source) : (int32_t)floor(*source);
   }
   else
     *dest = (int32_t) round(*source);
